CRM-21230 fix fatal error on multi-data imports
